requirements.txt
fastapi==0.115.0
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
chromadb==0.4.18
keybert==0.7.0
pydantic==2.5.2
jsonschema==4.19.1
ijson==3.2.3
